added Account and Company files
